+++ src/index.tsx
@@ -21,7 +21,9 @@ createRoot(document.getElementById('root') as HTMLDivElement).render(
       <Routes>
         <Route path="/" element={<App />}>
           <Route index element={<HomePage />} />
-          <Route path="people/:slugs?" element={<People />} />
+          <Route path="people" element={<People />}>
+            <Route path=":slugs" element={<People />} />
+          </Route>
           <Route path="*" element={<PageNotFound />} />
           <Route path="home" element={<Navigate to="/" replace />} />
         </Route>
